fix errors with one class training

+++ Dataset/TrainOneClassDataset.py
@@ -34,8 +34,8 @@ class TrainOneClassDataset(TrainDataset):
         random.Random(self.random_seed).shuffle(test_images_not_target)
         test_images_target = list(filter(lambda x: x[1] == 1, all_test_images))
         random.Random(self.random_seed).shuffle(test_images_target)
-        test_images_target = test_images_target[:self.target_image_percent * self.test_image_count]
-        test_images_not_target = test_images_not_target[:(1 - self.target_image_percent) * self.test_image_count]
+        test_images_target = test_images_target[:int(self.target_image_percent * self.test_image_count)]
+        test_images_not_target = test_images_not_target[:int((1 - self.target_image_percent) * self.test_image_count)]
         test_images_list = test_images_target + test_images_not_target
         test_paths, test_labels = self.unzip_list(test_images_list)
